Add version 0.01 of the index.html
Today, I created the most basic version of our project's Home page. In this commit, I added version 0.01 — a name I chose intentionally, without any technical significance. The content is an attempt to recreate the two "menus" found on the CPS pages: the government bar and the CPS bar (header). We encountered an issue with the search icon, which is an unknown problem at this point, and I’m not sure what caused it. But for today, this is more than enough. I hope we can make more progress in the coming days.